requirements.txt
Flask==3.0.2
numpy==1.26.1
keras==3.1.0
tensorflow==2.16.1
python-telegram-bot==21.0.1
Py
